requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
